Move template and runner classes into their own files
