+++ src/cfnlint/core.py
@@ -5,9 +5,11 @@ SPDX-License-Identifier: MIT-0
 import logging
 import os
 import sys
+
 from jsonschema.exceptions import ValidationError
 
-from cfnlint import Template
+import cfnlint.runner
+from cfnlint.template import Template
 from cfnlint.rules import RulesCollection
 import cfnlint.config
 import cfnlint.formatters
@@ -184,7 +186,7 @@ def run_checks(filename, template, rules, regions, mandatory_rules=None):
 
     matches = []
 
-    runner = cfnlint.Runner(rules, filename, template, regions, mandatory_rules=mandatory_rules)
+    runner = cfnlint.runner.Runner(rules, filename, template, regions, mandatory_rules=mandatory_rules)
     matches.extend(runner.transform())
     # Only do rule analysis if Transform was successful
     if not matches:
