+++ test/unit/module/test_template.py
@@ -6,7 +6,7 @@ import json
 import os
 from test.testlib.testcase import BaseTestCase
 import cfnlint.helpers
-from cfnlint import Template  # pylint: disable=E0401
+from cfnlint.template import Template  # pylint: disable=E0401
 
 
 class TestTemplate(BaseTestCase):
