+++ test/unit/rules/templates/test_yaml_template.py
@@ -3,7 +3,7 @@ Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: MIT-0
 """
 from test.unit.rules import BaseRuleTestCase
-from cfnlint import Runner  # pylint: disable=E0401
+from cfnlint.runner import Runner
 from cfnlint.rules import RulesCollection
 from cfnlint.rules.templates.Base import Base  # pylint: disable=E0401
